Refactor SSE, i18n, and layout handling with minor improvements
