+++ src/app/api/notifications/sse/route.ts
@@ -1,12 +1,10 @@
 import { NextRequest } from "next/server";
-
-// Simple in-memory broadcaster (per server instance)
-const clients = new Set<WritableStreamDefaultWriter<string>>();
+import { addClient, removeClient } from "@/lib/sse-utils";
 
 export async function GET(_req: NextRequest) {
   const stream = new TransformStream();
   const writer = stream.writable.getWriter();
-  clients.add(writer);
+  addClient(writer);
   const encoder = new TextEncoder();
 
   // Send initial ping
@@ -30,19 +28,19 @@ export async function GET(_req: NextRequest) {
   (response as any).finally?.(() => {
     clearInterval(interval);
     writer.close().catch(() => {});
-    clients.delete(writer);
+    removeClient(writer);
   });
 
   return response;
 }
 
-// Util for other routes to emit event
-export function emitOrderCreated(payload: any) {
-  const data = typeof payload === "string" ? payload : JSON.stringify(payload);
-  const encoder = new TextEncoder();
-  clients.forEach(async (w) => {
-    try {
-      await w.write(encoder.encode(`event: order\ndata: ${data}\n\n`) as any);
-    } catch {}
-  });
-}
+// Util for other routes to emit event - moved to separate utility file
+// export function emitOrderCreated(payload: any) {
+//   const data = typeof payload === "string" ? payload : JSON.stringify(payload);
+//   const encoder = new TextEncoder();
+//   clients.forEach(async (w) => {
+//     try {
+//       await w.write(encoder.encode(`event: order\ndata: ${data}\n\n`) as any);
+//     } catch {}
+//   });
+// }
